Pull request 642: Unpin from BoringSSL
Merge in ADGUARD-CORE-LIBS/dns-libs from fix/mips to master

Squashed commit of the following:

commit 62b13fe2d9fe689d030c8caed011d843f947ff49
Author: Sergey Fionov <fionov@gmail.com>
Date:   Tue Mar 12 11:12:57 2024 +0300

    Fix build

commit 54e160a68989a59ae1c3f271525ca4ec01afcf68
Author: Sergey Fionov <fionov@gmail.com>
Date:   Tue Mar 12 09:17:42 2024 +0300

    Bump

commit 25c7fceda55103fbf3ae222fd0b65dd9bd6ca093
Author: Sergey Fionov <fionov@gmail.com>
Date:   Tue Mar 12 08:57:29 2024 +0300

    Unpin from BoringSSL

+++ net/application_verifier.cpp
@@ -34,7 +34,7 @@ std::optional<std::string> ApplicationVerifier::verify(
 
     STACK_OF(X509) *chain = X509_STORE_CTX_get0_untrusted(ctx);
     event.chain.reserve(sk_X509_num(chain));
-    for (size_t i = 0; i < sk_X509_num(chain); ++i) {
+    for (size_t i = 0; i < size_t(sk_X509_num(chain)); ++i) {
         X509 *cert = sk_X509_value(chain, i);
         serialized = serialize_certificate(cert);
         if (serialized.has_value()) {
